build: changed total to a Decimal field with 2 decimal places and 10 …

+++ currencykidzapi/models/deposit_event.py
@@ -6,5 +6,5 @@ class DepositEvent(models.Model):
     name = models.CharField(max_length=50)
     saver = models.ForeignKey("Saver", on_delete=models.CASCADE)
     currency = models.ForeignKey("Currency", on_delete=models.CASCADE)
-    total = models.IntegerField()
+    total = models.DecimalField(max_digits=10 ,decimal_places=2)
     sound_effect = models.URLField()
